[#70 & #68] refactor: change package naming for bootsrapper
